move test-util to src
