+++ packages/gen-ai-hub/src/client/openai/openai-client.test.ts
@@ -1,6 +1,6 @@
 import nock from 'nock';
 import { HttpDestination } from '@sap-cloud-sdk/connectivity';
-import { mockGetAiCoreDestination } from '../../../test-util/mock-context.js';
+import { mockGetAiCoreDestination } from '../../test-util/mock-context.js';
 import {
   BaseLlmParametersWithDeploymentId,
   EndpointOptions
@@ -8,7 +8,7 @@ import {
 import {
   mockInference,
   parseMockResponse
-} from '../../../test-util/mock-http.js';
+} from '../../test-util/mock-http.js';
 import { OpenAiClient } from './openai-client.js';
 import {
   OpenAiChatCompletionOutput,
