+++ packages/gen-ai-hub/src/orchestration/orchestration-client.test.ts
@@ -1,7 +1,7 @@
 import nock from 'nock';
 import { HttpDestination } from '@sap-cloud-sdk/connectivity';
-import { mockGetAiCoreDestination } from '../../test-util/mock-context.js';
-import { mockInference, parseMockResponse } from '../../test-util/mock-http.js';
+import { mockGetAiCoreDestination } from '../test-util/mock-context.js';
+import { mockInference, parseMockResponse } from '../test-util/mock-http.js';
 import { BaseLlmParametersWithDeploymentId } from '../core/index.js';
 import {
   GenAiHubClient,
