+++ packages/gen-ai-hub/src/test-util/mock-data/openai/openai-chat-completion-success-response.json
