+++ packages/gen-ai-hub/src/test-util/mock-data/openai/openai-embeddings-success-response.json
