+++ packages/gen-ai-hub/src/test-util/mock-data/openai/openai-error-response.json
