+++ packages/gen-ai-hub/src/test-util/mock-data/orchestration/genaihub-chat-completion-success-response.json
